Add integration tests

+++ PersonHubApi/tests/PersonHub.IntegrationTest/Fixtures/IntegrationTestClassFixture.cs
@@ -22,44 +22,48 @@ namespace PersonHub.IntegrationTest.Fixtures
 
         public IntegrationTestClassFixture()
         {
-            //TODO: Read configuration from appsettings instead of hardcode here
-            var connectionStringBuilder = new NpgsqlConnectionStringBuilder()
+            const bool shouldRunFlyway = false;
+            if (shouldRunFlyway)
             {
-                Host = "localhost",
-                Port = 5432,
-                Database = "person-hub-test",
-                Username = "postgres",
-                Password = "P@ssw0rd",
-                SearchPath = "public"
-            };
+                //TODO: Read configuration from appsettings instead of hardcode here
+                var connectionStringBuilder = new NpgsqlConnectionStringBuilder()
+                {
+                    Host = "localhost",
+                    Port = 5432,
+                    Database = "person-hub-test",
+                    Username = "postgres",
+                    Password = "P@ssw0rd",
+                    SearchPath = "public"
+                };
 
-            using (var connection = new NpgsqlConnection(connectionStringBuilder.ConnectionString))
-            {
-                connection.Execute(@"set Search_Path to ""public""; 
+                using (var connection = new NpgsqlConnection(connectionStringBuilder.ConnectionString))
+                {
+                    connection.Execute(@"set Search_Path to ""public""; 
                                             DROP SCHEMA IF EXISTS ""person-hub-test"" CASCADE; 
                                             CREATE SCHEMA ""person-hub-test"";
                                             grant usage on schema ""person-hub-test"" to public;
                                             grant create on schema ""person-hub-test"" to public;
                                             ");
-            }
-
-            connectionStringBuilder.SearchPath = "person-hub-test";
-            using (var connection = new NpgsqlConnection(connectionStringBuilder.ConnectionString))
-            {
-                //TODO: Should found a robust way to get sql path
-                string path = Path.Combine(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location), @"../../../../../sql");
-                var files = Directory.GetFiles(path).OrderBy(r => r);
-                var setSearchPathSql = "set Search_Path to \"person-hub-test\";";
-
-                if (!files.Any())
-                {
-                    throw new Exception("WRONG SQL PATH FOLDER");
                 }
 
-                foreach (var filePath in files)
+                connectionStringBuilder.SearchPath = "person-hub-test";
+                using (var connection = new NpgsqlConnection(connectionStringBuilder.ConnectionString))
                 {
-                    string sqlContent = File.ReadAllText(filePath);
-                    connection.Execute(setSearchPathSql + sqlContent);
+                    //TODO: Should found a robust way to get sql path
+                    string path = Path.Combine(Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location), @"../../../../../sql");
+                    var files = Directory.GetFiles(path).OrderBy(r => r);
+                    var setSearchPathSql = "set Search_Path to \"person-hub-test\";";
+
+                    if (!files.Any())
+                    {
+                        throw new Exception("WRONG SQL PATH FOLDER");
+                    }
+
+                    foreach (var filePath in files)
+                    {
+                        string sqlContent = File.ReadAllText(filePath);
+                        connection.Execute(setSearchPathSql + sqlContent);
+                    }
                 }
             }
